requirements.txt
# ── Core ──
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
websockets>=12.0
pydantic[email]>=2.0
pydantic-settings>=2.0

# ── Database ──
sqlalchemy[asyncio]>=2.0
asyncpg>=0.30.0
alembic>=1.14.0
azure-cosmos>=4.9.0
aiohttp>=3.9.0

# ── Cache ──
redis>=5.0

# ── Auth ──
PyJWT>=2.9
bcrypt>=4.2

# ── Azure Communication Services (Email) ──
azure-communication-email>=1.0.0

# ── Azure AI (AzureOpenAI + Responses API) ──
openai>=1.70.0
azure-identity>=1.19.0

# ── Azure Speech (REST API for voice pipeline) ──
httpx>=0.27.0

# ── MCP Server (Korean corpus tools, mounted at /mcp) ──
fastmcp>=3.1.0

# ── Utilities ──
python-dotenv>=1.0.0
